Migrate GasCalculator to shadcn registry (#14)
- Update GasCalculator to use components from shadcn registry
- Add components.json for shadcn configuration
- Add src/lib/utils.ts for cn utility
- Update installation instructions to use shadcn CLI
- Remove old local component files
