+++ src/app/docs/components/gas-calculator/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState } from "react";
-import { GasCalculator } from "./component";
+import { GasCalculator } from "@/components/w3-kit/gas-calculator";
 import { Code, Eye } from "lucide-react";
 import { CodeBlock } from "@/components/docs/codeBlock";
 
@@ -117,7 +117,7 @@ export default function Page() {
                   <p className="text-sm text-gray-600 dark:text-gray-400 mb-4">
                     Run the following command to add the Gas Calculator component to your project:
                   </p>
-                  <CodeBlock code="npx w3-kit@latest add gas-calculator" id="cli" />
+                  <CodeBlock code="npx shadcn@latest add https://w3-kit.com/registry/gas-calculator.json" id="cli" />
                   <p className="text-sm text-gray-600 dark:text-gray-400 mt-4">
                     This will:
                   </p>
